technodrome/requirements.txt
certifi==2020.12.5
chardet==4.0.0
Django==3.1.5
djangorestframework==3.12.2
influxdb==5.3.1
psycopg2==2.8.6
python-dateutil== 2.8.1
pytz==2020.5
requests==2.25.1
six==1.15.0
urllib3==1.26.3
paho-mqtt==1.5.1
attrs==20.3.0